wrapped object in array to conform to strange antd
