+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lumi-cli",
-  "version": "0.0.34",
+  "version": "0.0.35",
   "description": "",
   "bin": {
     "lumi": "./bin/index"
